refactor: auth module structure
