+++ src/modules/auth/auth.controller.ts
@@ -1,8 +1,8 @@
 import { Controller, Req, Post, UseGuards, Get } from '@nestjs/common';
 import { Request } from 'express';
-import { LocalAuthGuard } from './local-auth.guard';
+import { LocalAuthGuard } from './guards/local-auth.guard';
 import { AuthService } from './auth.service';
-import { JwtAuthGuard } from './jwt-auth.guard';
+import { JwtAuthGuard } from './guards/jwt-auth.guard';
 
 @Controller('authentications')
 export class AuthController {
